+++ src/go/cmd/testdata/test_external_alias_source/test.go
@@ -1,0 +1,5 @@
+package test_external_alias_source
+
+type Foo struct{
+	Bar string
+}