+++ src/go/cmd/testdata/test_service_group/group/internal/test.go
@@ -1,0 +1,5 @@
+package internal
+
+type Foo struct{
+	Bar string
+}