+++ tools/sdk-generation-pipeline/packages/sdk-generation-cli/src/cli/dockerCli/core/DockerRunningModel.ts
@@ -1,0 +1,5 @@
+export enum DockerRunningModel {
+    CodeGenAndGrowUp,
+    GrowUp,
+    Pipeline
+}