+++ tools/sdk-testgen/packages/autorest.gotest/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @autorest/gotest
 
-This log was last generated on Mon, 25 Apr 2022 08:06:55 GMT and should not be manually modified.
+This log was last generated on Mon, 23 May 2022 07:05:43 GMT and should not be manually modified.
+
+## 4.0.1
+Mon, 23 May 2022 07:05:43 GMT
+
+### Patches
+
+- Fix module import problem when SDK version bigger than v1.
+
+## 4.0.0
+Mon, 16 May 2022 01:46:35 GMT
+
+### Breaking changes
+
+- Align test code with GA core lib.
 
 ## 3.1.2
 Mon, 25 Apr 2022 08:06:55 GMT